robotic_sorting_system/requirements.txt
ultralytics
opencv-python
pyserial
Pillow
